Fix build issues
