+++ packages/swap/src/helpers/order.ts
@@ -1,7 +1,4 @@
 import {Balance, Swap} from '@yoroi/types'
-import {BalanceQuantity} from '@yoroi/types/lib/balance/token'
-import {Quantities} from '@yoroi/wallet-mobile/src/yoroi-wallets/utils'
-import BigNumber from 'bignumber.js'
 
 type AmountPair = {
   sell: Balance.Amount
@@ -92,38 +89,6 @@ export const getSellAmountByChangingReceive = (
         }
 
   return {buy, sell}
-}
-
-export const getBuyQuantityForLimitOrder = (
-  sellQuantityDenominated: BalanceQuantity,
-  limitPrice: BalanceQuantity,
-  buyTokenDecimals: number,
-): BalanceQuantity => {
-  if (Quantities.isZero(limitPrice)) {
-    return Quantities.zero
-  }
-
-  return Quantities.integer(
-    Quantities.quotient(sellQuantityDenominated, limitPrice),
-    buyTokenDecimals,
-  ).toString() as BalanceQuantity
-}
-
-export const getSellQuantityForLimitOrder = (
-  buyQuantityDenominated: BalanceQuantity,
-  limitPrice: BalanceQuantity,
-  sellTokenDecimals: number,
-): BalanceQuantity => {
-  if (Quantities.isZero(limitPrice)) {
-    return Quantities.zero
-  }
-
-  return Quantities.integer(
-    BigNumber(buyQuantityDenominated)
-      .times(BigNumber(limitPrice))
-      .toString() as BalanceQuantity,
-    sellTokenDecimals,
-  ).toString() as BalanceQuantity
 }
 
 export const makeLimitOrder = (
